requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
requests
passlib
bcrypt
python-jose
